srv_server/requirements.txt
grpcio==1.43.0
grpcio-tools==1.43.0
psycopg2-binary==2.9.9
redis==3.5.3